+++ client/components/all-product-container.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import SingleProductView from './all-product-single-view'
-import {fetchAllProducts} from '../store/product'
 import {Grid} from 'semantic-ui-react'
+import SingleProductView from './index'
+import {fetchAllProducts} from '../store'
 
 class AllProducts extends React.Component {
   componentDidMount() {
