+++ client/components/category-bar.js
@@ -1,6 +1,5 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
-import {Link} from 'react-router-dom'
 import {Menu} from 'semantic-ui-react'
 import {fetchAllCategories, fetchAllProducts} from '../store'
