+++ client/components/index.js
@@ -14,4 +14,6 @@ export {default as Checkout} from './checkout'
 export {default as ShippingAddressForm} from './shipping-address-form'
 export {default as OrderProducts} from './order-products'
 export {default as OrderConfirmation} from './order-confirmation'
-export {default as CategoryBar} from './categorybar'
+export {default as CategoryBar} from './category-bar'
+export {default as CheckoutForm} from './stripe-components/checkout-form'
+export {default as StripeContainer} from './stripe-components/stripe-container'
