+++ server/api/users.js
@@ -7,9 +7,7 @@ router.get('/', async (req, res, next) => {
     if (req.user) {
       if (req.user.isAdmin) {
         const users = await User.findAll({
-          // explicitly select only the id and email fields - even though
-          // users' passwords are encrypted, it won't help if we just
-          // send everything to anyone who asks!
+          // explicitly select only the id and email fields
           attributes: ['id', 'email']
         })
         res.json(users)
